job done

+++ Domain Layer/DbModels/RolePermission.cs
@@ -4,7 +4,7 @@ public class RolePermission
 {
     public int Id { get; set; }
     public int PermissionId { get; set; }
-    public int RoleId { get; set; }
+    public long RoleId { get; set; }
     public DateTime CreatedDate { get; set; }
     public string CreatedBy { get; set; }
