+++ Domain Layer/DbModels/UserRole.cs
@@ -3,8 +3,8 @@ namespace Domain_Layer.DbModels;
 public class UserRole
 {
     public long Id { get; set; }
-    public long UserAccountId { get; set; }
-    public int RoleId { get; set; }
+    public int UserAccountId { get; set; }
+    public long RoleId { get; set; }
     public DateTime CreatedDate { get; set; }
     public string CreatedBy { get; set; }
